Create Stream message and docker-compose with RabbitMQ and reactive REst API

+++ microservices/recommendation-service/src/main/java/com/jfcdevs/app/services/RecommendationServiceImpl.java
@@ -10,9 +10,11 @@ import com.mongodb.DuplicateKeyException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
-import java.util.ArrayList;
-import java.util.List;
+
+import static java.util.logging.Level.*;
 
 @RestController
 @Slf4j
@@ -28,34 +30,46 @@ public class RecommendationServiceImpl implements RecommendationService {
         this.mapper = mapper;
     }
     @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
 
         if(productId < 1){
             throw new InvalidInputException("Invalid productId: " + productId);
         }
-        List<RecommendationEntity> entityList = repository.findByProductId(productId);
-        List<Recommendation> list = mapper.entityListToApiList(entityList);
-        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
-        log.debug("getRecommendation: response size: {}", list.size());
-        return list;
+
+        log.debug("getRecommendation: response for productId {}", productId);
+        return repository.findByProductId(productId)
+                .log(log.getName(), FINE)
+                .map(e -> mapper.entityToApi(e))
+                .map(e -> setServiceAddress(e));
     }
 
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
-        try{
-            RecommendationEntity entity = mapper.apiToEntity(body);
-            RecommendationEntity newEntity = repository.save(entity);
-
-            log.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(), body.getRecommendationId());
-            return mapper.entityToApi(newEntity);
-        }catch (DuplicateKeyException dke){
-            throw new InvalidInputException("Duplicate key, productId: "+body.getProductId()+" recommendationId: "+body.getRecommendationId());
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
+        if(body.getProductId() < 1){
+            throw new InvalidInputException("Invalid productId: " + body.getProductId());
         }
+        RecommendationEntity entity = mapper.apiToEntity(body);
+        Mono<Recommendation> newEntity = repository.save(entity)
+                .log(log.getName(), FINE)
+                .onErrorMap(
+                        DuplicateKeyException.class,
+                        ex -> new InvalidInputException("Duplicate key, Product Id: "+  body.getProductId()+ ", Recommendation Id: "+body.getRecommendationId())
+                )
+                .map(e -> mapper.entityToApi(e));
+        return newEntity;
     }
 
     @Override
-    public void deleteRecommendation(int productId) {
-        log.debug("deleteRecommendation: tries to delete recommendatioons for the product with productId: {}", productId);
-        repository.deleteAll(repository.findByProductId(productId));
+    public Mono<Void> deleteRecommendation(int productId) {
+        if(productId < 1){
+            throw new InvalidInputException("Invalid productId: " + productId);
+        }
+        log.debug("deleteRecommendation: tries to delete recommendations for the product with productId: {}", productId);
+        return repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private Recommendation setServiceAddress(Recommendation e){
+        e.setServiceAddress(serviceUtil.getServiceAddress());
+        return e;
     }
 }
